Position name holder and recenter button
