--- client/src/component/avatarpicker/styles.css
@@ -1,4 +1,0 @@
-.appear-on-left {
-  transform: translate(300px, 50px) !important;
-  width: fit-content;
-}
