--- client/src/component/nameholder/styles.css
@@ -1,3 +1,0 @@
-.row-flex {
-  display: flex;
-}
