--- client/src/component/picturecontainer/imagesList.js
@@ -1,15 +1,0 @@
-import boy1 from '../avatarpicker/images/boy.svg';
-import boy2 from '../avatarpicker/images/bussiness-man.svg';
-import boy3 from '../avatarpicker/images/man.svg';
-import girl1 from '../avatarpicker/images/girl.svg';
-import girl2 from '../avatarpicker/images/girl2.svg';
-import girl3 from '../avatarpicker/images/woman.svg';
-
-export const listOfImages = {
-  'boy1': boy1,
-  'boy2': boy2,
-  'boy3': boy3,
-  'girl1': girl1,
-  'girl2': girl2,
-  'girl3': girl3,
-}
