+++ client/src/components/picturecontainer/imagesList.js
@@ -1,0 +1,15 @@
+import boy1 from '../images/boy.svg';
+import boy2 from '../images/bussiness-man.svg';
+import boy3 from '../images/man.svg';
+import girl1 from '../images/girl.svg';
+import girl2 from '../images/girl2.svg';
+import girl3 from '../images/woman.svg';
+
+export const listOfImages = {
+  boy1: boy1,
+  boy2: boy2,
+  boy3: boy3,
+  girl1: girl1,
+  girl2: girl2,
+  girl3: girl3,
+};
