+++ client/src/components/picturecontainer/styles.css
@@ -2,15 +2,15 @@
   background-color: #e5e5e5;
   border-radius: 40px;
   display: grid;
-  grid-template-columns: auto auto auto;
   gap: 16px;
-  padding: 16px 5px;
+  grid-template-columns: auto auto auto;
   justify-content: center;
+  padding: 16px 5px;
 }
 
 .image-style {
-  border-radius: 40px;
   border: 3px solid transparent;
+  border-radius: 40px;
   height: auto;
   width: 70px;
 }
